[EMCAL-610] Fix segfault in Raw creator
- Fix buffer overrun in findBunches leading to segfault
- In getting online ID (SRU ID in supermodule had to be
   scaled with the number of supermodules in order to get
   the DDL ID)
- Convert trailer size as size in 32bit words to size in byte in
   carryOverMethod in order to calculate properly the
   payload size

+++ Detectors/EMCAL/base/src/Mapper.cxx
@@ -115,6 +115,6 @@ Mapper& MappingHandler::getMappingForDDL(int ddl)
 
 std::ostream& o2::emcal::operator<<(std::ostream& stream, const Mapper::ChannelID& channel)
 {
-  stream << "Row " << channel.mRow << ", Column " << channel.mColumn << ", type " << o2::emcal::channelTypeToString(channel.mChannelType);
+  stream << "Row " << static_cast<int>(channel.mRow) << ", Column " << static_cast<int>(channel.mColumn) << ", type " << o2::emcal::channelTypeToString(channel.mChannelType);
   return stream;
 }
